refactor gameActionResolver to improve readability
