+++ src/reducers/gameReducer/resolver/addCow/index.ts
@@ -1,0 +1,5 @@
+import canAddCow from './canAddCow';
+import canAddCowResolver from './canAddCowResolver';
+
+export { canAddCow };
+export default canAddCowResolver;
