+++ src/reducers/gameReducer/resolver/gameActionResolver.ts
@@ -1,9 +1,5 @@
-import {
-  AddCowPayload, AddCowResolvedAction, GameActionResolver,
-  GameActionTypeName,
-  GamePayloadTypeName,
-  GameStatus
-} from '../../../utils/types';
+import { GameActionResolver } from '../../../utils/types';
+import canAddCowResolver, { canAddCow } from './addCow';
 import { isCowId, isPadId } from './isResourceId';
 
 const gameActionResolver: GameActionResolver = (selectableId, game) => {
@@ -11,26 +7,7 @@ const gameActionResolver: GameActionResolver = (selectableId, game) => {
   }
 
   if (isPadId(selectableId)) {
-    const canAddCow =
-      game.actionStack.length === 0 &&
-      game.cows.safeCows[game.currentPlayer] > 0 &&
-      game.gameStatus === GameStatus.ONGOING &&
-      game.pads[selectableId].visitingCow === null;
-
-    // ADD Cow
-    if (canAddCow) {
-      const addCowPayload: AddCowPayload = {
-        __typename: GamePayloadTypeName.ADD_COW_PAYLOAD,
-        selectedPadId: selectableId,
-      };
-      const addCowResolvedAction: AddCowResolvedAction = {
-        __typename: GameActionTypeName.ADD_COW_ACTION,
-        type: GameActionTypeName.ADD_COW_ACTION,
-        payload: addCowPayload,
-      };
-
-      return addCowResolvedAction;
-    }
+    if (canAddCow(selectableId, game)) return canAddCowResolver(selectableId);
   }
 
   return null;
